Put idempotency tests in their own category
The vulpix command will not run them anymore.
